+++ services/ui/src/js/components/Chart/ChartDensity.js
@@ -1,17 +1,16 @@
+import ChartElement from "./ChartElement";
+
 /**
  * Contour density plot web component.
  *
  * Useful for displaying a two-dimensional distribution, such as those found in
  * vector variables like wind.
  */
-export default class ChartDensity extends HTMLElement {
+export default class ChartDensity extends ChartElement {
   static #TEMPLATE = `<p>Coming soon!</p>`;
   static #STYLE = ``;
 
-  constructor() {
-    super();
-
-    const root = this.attachShadow({ mode: "open" });
-    root.innerHTML = ChartDensity.#STYLE + ChartDensity.#TEMPLATE;
+  static get observedAttributes() {
+    return ChartElement.observedAttributes;
   }
 }
