+++ services/ui/src/js/components/Chart/ChartHistogram.js
@@ -12,70 +12,17 @@ import {
   select,
 } from "d3";
 
-class Histogram extends HTMLElement {
-  static #TEMPLATE = `<span id="title-y"></span>
-  <svg>
-    <g class="x-axis"></g>
-    <g class="y-axis"></g>
-    <g class="data"></g>
-    <g class="annotation"></g>
-  </svg>
-  <span id="title-x"></span>`;
+import ChartElement from "./ChartElement";
 
-  static #STYLE = `<style>
-  :host {
-    display: grid;
-    gap: 0.5em;
-    grid-template-columns: min-content 1fr;
-    grid-template-rows: 1fr min-content;
-    place-items: center;
-  }
-
-  #title-x,
-  svg {
-    grid-column: 2 / 3;
-  }
-
-  #title-y,
-  svg {
-    grid-row: 1 / 2;
-  }
-
-  #title-y {
-    writing-mode: vertical-lr;
-    transform: rotate(180deg);
-  }
-
-  #title-x,
-  .x-axis {
-    color: #3d4551;
-  }
-
-  #title-y,
-  .y-axis {
-    color: #71767a;
-  }
-
-  svg {
-    place-self: stretch;
-    aspect-ratio: 4 / 3;
-  }
-
-  line,
-  path,
-  rect {
-    shape-rendering: crispEdges;
-  }
-
-  .deviation rect {
+class ChartHistogram extends ChartElement {
+  static #STYLE = `.deviation rect {
     fill: #dfe1e2;
     mix-blend-mode: color-burn;
   }
 
   .annotation line {
     stroke: currentColor;
-  }
-  </style>`;
+  }`;
 
   #data = [];
   #thresholds = null;
@@ -83,53 +30,12 @@ class Histogram extends HTMLElement {
   #mean = 0;
   #deviation = 0;
 
-  constructor() {
-    super();
-
-    const shadowRoot = this.attachShadow({ mode: "open" });
-    shadowRoot.innerHTML = Histogram.#STYLE + Histogram.#TEMPLATE;
-  }
-
   static get observedAttributes() {
-    return ["format-x", "format-y", "src", "title-x", "title-y"];
+    return ChartElement.observedAttributes;
   }
 
-  connectedCallback() {
-    this.resizeObserver = new ResizeObserver(() => {
-      this.render();
-    });
-    this.resizeObserver.observe(this.shadowRoot?.querySelector("svg"));
-  }
-
-  disconnectedCallback() {
-    this?.resizeObserver.unobserve(this.shadowRoot?.querySelector("svg"));
-    delete this?.resizeObserver;
-  }
-
-  attributeChangedCallback(name, oldValue, newValue) {
-    switch (name) {
-      case "src":
-        fetch(newValue)
-          .then((response) => response.json())
-          .then((data) => {
-            this.data = data;
-            this.dispatchEvent(new CustomEvent("data-load", { bubbles: true }));
-          })
-          .catch((err) => {
-            console.error(err);
-          });
-        break;
-      case "format-x":
-      case "format-y":
-        this.render();
-        break;
-      case "title-x":
-      case "title-y":
-        this.#updateLabel(name, newValue);
-        break;
-      default:
-        break;
-    }
+  get componentStyles() {
+    return super.componentStyles + ChartHistogram.#STYLE;
   }
 
   get bins() {
@@ -157,30 +63,6 @@ class Histogram extends HTMLElement {
     this.render();
   }
 
-  get formatX() {
-    return this.getAttribute("format-x") ?? ",";
-  }
-
-  set formatX(value) {
-    if (!value) {
-      this.removeAttribute("format-x");
-    } else {
-      this.setAttribute("format-x", value);
-    }
-  }
-
-  get formatY() {
-    return this.getAttribute("format-y") ?? ",";
-  }
-
-  set formatY(value) {
-    if (!value) {
-      this.removeAttribute("format-y");
-    } else {
-      this.setAttribute("format-y", value);
-    }
-  }
-
   get range() {
     return this.#range;
   }
@@ -190,19 +72,6 @@ class Histogram extends HTMLElement {
     this.render();
   }
 
-  get src() {
-    return this.getAttribute("src");
-  }
-
-  set src(value) {
-    if (!value) {
-      this.removeAttribute("src");
-      this.data = [];
-    } else {
-      this.setAttribute("src", value);
-    }
-  }
-
   get thresholds() {
     return this.#thresholds;
   }
@@ -210,30 +79,6 @@ class Histogram extends HTMLElement {
   set thresholds(value) {
     this.#thresholds = value;
     this.render();
-  }
-
-  get titleX() {
-    return this.getAttribute("title-x");
-  }
-
-  set titleX(value) {
-    if (!value) {
-      this.removeAttribute("title-x");
-    } else {
-      this.setAttribute("title-x", value);
-    }
-  }
-
-  get titleY() {
-    return this.getAttribute("title-y");
-  }
-
-  set titleY(value) {
-    if (!value) {
-      this.removeAttribute("title-y");
-    } else {
-      this.setAttribute("title-y", value);
-    }
   }
 
   render() {
@@ -358,12 +203,6 @@ class Histogram extends HTMLElement {
         g.selectAll(".tick text").attr("x", 4).attr("dy", -4);
       });
   }
-
-  #updateLabel(id, value) {
-    const span = this.shadowRoot?.getElementById(id);
-    if (!span) return;
-    span.textContent = value;
-  }
 }
 
-export default Histogram;
+export default ChartHistogram;
