+++ services/ui/src/js/components/Chart/index.js
@@ -1,0 +1,2 @@
+export { default as ChartDensity } from "./ChartDensity";
+export { default as ChartHistogram } from "./ChartHistogram";
