--- services/ui/src/js/components/ChartDensity/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ChartDensity";
