+++ services/ui/src/js/main.js
@@ -5,11 +5,10 @@ import styles from "../styles/_index.scss"; // eslint-disable-line
 import App from "./components/App";
 import DataSource from "./components/DataSource";
 import DiagScalar from "./components/DiagScalar";
-import Histogram from "./components/Histogram";
-import ChartDensity from "./components/ChartDensity";
+import { ChartDensity, ChartHistogram } from "./components/Chart";
 import ChartMap from "./components/ChartMap";
 
-customElements.define("chart-histogram", Histogram);
+customElements.define("chart-histogram", ChartHistogram);
 customElements.define("chart-density", ChartDensity);
 customElements.define("chart-map", ChartMap);
 customElements.define("data-source", DataSource);
